resolved texture leakage and optimizers inherited from optimizer.ts
