+++ src/session_test.ts
@@ -20,7 +20,7 @@ import {NDArrayMathGPU} from './math/math_gpu';
 import {Array1D, NDArray, Scalar} from './math/ndarray';
 import {FeedDictionary, FeedEntry, Session} from './session';
 import {SGDOptimizer} from './sgd_optimizer';
-import {MomentumOptimizer} from './momentumOptimizer';
+import {MomentumOptimizer} from './momentum_optimizer';
 import {AdagradOptimizer} from './adagrad_optimizer';
 import {RmspropOptimizer} from './rmsprop_optimizer';
