+++ src/sgd_optimizer.ts
@@ -15,46 +15,13 @@ limitations under the License.
 
 import {Node} from './graph';
 import {NDArrayMath} from './math/math';
-import {NDArray, Scalar} from './math/ndarray';
 import {Optimizer} from './optimizer';
 import {SessionRuntime} from './session';
-import * as session_util from './session_util';
 import {TensorArrayMap, SummedTensorArrayMap} from './tensor_array_map';
 
 export class SGDOptimizer extends Optimizer {
   constructor(protected learningRate: number, specifiedVariableList?: Node[]) {
-    super(specifiedVariableList);
-  }
-
-  beforeBatch(
-      math: NDArrayMath, batchSize: number, runtime: SessionRuntime,
-      activationArrayMap: TensorArrayMap,
-      gradientArrayMap: SummedTensorArrayMap) {
-    this.variableNodes = this.specifiedVariableNodes == null ?
-        session_util.getVariableNodesFromEvaluationSet(runtime.nodes) :
-        this.specifiedVariableNodes;
-    if (batchSize !== this.prevBatchSize) {
-      this.prevBatchSize = batchSize;
-      this.c = Scalar.new(-this.learningRate / batchSize);
-    }
-    this.variableNodes.forEach(
-        node => this.variableGradients.set(
-            node.output, NDArray.zeros(node.output.shape)));
-  }
-
-  afterExample(
-      math: NDArrayMath, runtime: SessionRuntime,
-      activationArrayMap: TensorArrayMap,
-      gradientArrayMap: SummedTensorArrayMap) {
-    math.scope((keep) => {
-      this.variableNodes.forEach(node => {
-        const gradient = gradientArrayMap.get(node.output);
-        const accumulatedGradient = this.variableGradients.get(node.output);
-        this.variableGradients.set(
-            node.output, keep(math.add(gradient, accumulatedGradient)));
-        accumulatedGradient.dispose();
-      });
-    });
+    super(learningRate, specifiedVariableList);
   }
 
   afterBatch(
@@ -79,18 +46,10 @@ export class SGDOptimizer extends Optimizer {
   }
 
   dispose() {
-    if (this.c != null) {
-      this.c.dispose();
-    }
-    this.one.dispose();
+    super.dispose();
   }
 
   setLearningRate(learningRate: number) {
     this.learningRate = learningRate;
   }
-
-  protected variableGradients = new TensorArrayMap();
-  protected prevBatchSize: number;
-  protected one = Scalar.new(1);
-  protected c: Scalar;
 }
